AArch64 task startup code
has to arrange for r30 to be the pointer to the task stop function once
the given task pc has been jumped to, hence the little bit of
trampoline code. AArch64 caches the last return address in r30, so leaf
code does not need to push to memory and ret just does a jump to r30.