should be ui !
Also need to think about how to handle reference counts of environments
captured for use as objects. So, (defun new () (env)), will allow you
to say (defq a (new)) etc. There is a conflict with UI trees wanting no
reference to the parent hash_map and yet this idea would want a
reference to the parent…..mmmmm