don't use quasi-quotation where it's obviously overkill
it’s tricky to gauge if this is quicker or not. In some situations it’s
obvious it must be, but in others it’s not so easy. There’s not a huge
amount in it either way so I’ll go on what code looks clearer to my eye.