free with mail free call
this is going to matter from now on as the headers will become split
from the data buffers, which will be reference counted immutable string
objects. msg headers will be more like stamps that point to data to be
sent. A big parcel will be split into multiple sections each with a
stamp addressing fragments of it.